Update patch set 1
Patch Set 1:

> Looks good to me.
 > One Q, why the Log statement in else block was removed?

This was one place I wasn't sure whether or not to include it, since in some cases it's expected. If we're doing a GET-NEXT, and there isn't any next object in the tree, null is a valid response; Or in the case of a GET-WALK, where we send out consecutive GET-BULK requests until we get back null.

Patch-set: 1
Reviewer: Gerrit User 3285 <3285@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0